split test.po into different files and made sorting in the main directory
